Bug Fix for Adult Freesurfer workflow
